[Oztechan/CCC#1821] Relocate common and client utils and create constants
