+++ client/src/commonMain/kotlin/com/oztechan/ccc/client/mapper/ExchangeRate.kt
@@ -1,7 +1,7 @@
 package com.oztechan.ccc.client.mapper
 
+import com.oztechan.ccc.client.util.nowAsInstant
 import com.oztechan.ccc.client.util.toDateString
-import com.oztechan.ccc.common.core.infrastructure.util.nowAsInstant
 import com.oztechan.ccc.common.core.model.ExchangeRate
 
 internal fun ExchangeRate.toConversion() = conversion.copy(base = base, date = nowAsInstant().toDateString())
