+++ client/src/commonMain/kotlin/com/oztechan/ccc/client/util/PremiumUtil.kt
@@ -1,0 +1,47 @@
+package com.oztechan.ccc.client.util
+
+import com.oztechan.ccc.client.model.PremiumType
+import kotlinx.datetime.DateTimeUnit
+import kotlinx.datetime.TimeZone
+import kotlinx.datetime.plus
+
+internal const val REWARDED_AD_PREMIUM_IN_DAYS = 2
+
+@Suppress("MagicNumber")
+internal fun PremiumType.calculatePremiumEnd(startDate: Long = nowAsLong()) = when (this) {
+    PremiumType.VIDEO -> startDate.toInstant().plus(
+        REWARDED_AD_PREMIUM_IN_DAYS,
+        DateTimeUnit.DAY,
+        TimeZone.currentSystemDefault()
+    ).toEpochMilliseconds()
+
+    PremiumType.MONTH -> startDate.toInstant().plus(
+        1,
+        DateTimeUnit.MONTH,
+        TimeZone.currentSystemDefault()
+    ).toEpochMilliseconds()
+
+    PremiumType.QUARTER -> startDate.toInstant().plus(
+        3,
+        DateTimeUnit.MONTH,
+        TimeZone.currentSystemDefault()
+    ).toEpochMilliseconds()
+
+    PremiumType.HALF_YEAR -> startDate.toInstant().plus(
+        6,
+        DateTimeUnit.MONTH,
+        TimeZone.currentSystemDefault()
+    ).toEpochMilliseconds()
+
+    PremiumType.YEAR -> startDate.toInstant().plus(
+        1,
+        DateTimeUnit.YEAR,
+        TimeZone.currentSystemDefault()
+    ).toEpochMilliseconds()
+
+    PremiumType.LIFE_TIME -> startDate.toInstant().plus(
+        1,
+        DateTimeUnit.CENTURY,
+        TimeZone.currentSystemDefault()
+    ).toEpochMilliseconds()
+}
