+++ client/src/commonTest/kotlin/com/oztechan/ccc/client/repository/AdRepositoryTest.kt
@@ -6,8 +6,8 @@ import com.oztechan.ccc.client.helper.BaseSubjectTest
 import com.oztechan.ccc.client.repository.ad.AdRepository
 import com.oztechan.ccc.client.repository.ad.AdRepositoryImpl
 import com.oztechan.ccc.client.storage.app.AppStorage
-import com.oztechan.ccc.common.core.infrastructure.util.SECOND
-import com.oztechan.ccc.common.core.infrastructure.util.nowAsLong
+import com.oztechan.ccc.client.util.nowAsLong
+import com.oztechan.ccc.common.core.infrastructure.constants.SECOND
 import io.mockative.Mock
 import io.mockative.classOf
 import io.mockative.given
