+++ client/src/commonTest/kotlin/com/oztechan/ccc/client/util/DateUtilTest.kt
@@ -1,26 +1,34 @@
 package com.oztechan.ccc.client.util
 
 import com.oztechan.ccc.client.helper.BaseTest
-import com.oztechan.ccc.client.model.PremiumType
-import com.oztechan.ccc.common.core.infrastructure.util.DAY
-import com.oztechan.ccc.common.core.infrastructure.util.SECOND
-import com.oztechan.ccc.common.core.infrastructure.util.nowAsLong
-import kotlinx.datetime.DateTimePeriod
+import com.oztechan.ccc.common.core.infrastructure.constants.DAY
+import com.oztechan.ccc.common.core.infrastructure.constants.SECOND
+import kotlinx.datetime.Clock
 import kotlinx.datetime.Instant
 import kotlinx.datetime.TimeZone
-import kotlinx.datetime.plus
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 
 internal class DateUtilTest : BaseTest() {
+
     @Test
-    fun isPremiumExpired() {
-        assertTrue { (nowAsLong() - DAY).isPremiumExpired() }
-        assertTrue { (nowAsLong() - SECOND).isPremiumExpired() }
-        assertFalse { (nowAsLong() + DAY).isPremiumExpired() }
-        assertFalse { (nowAsLong() + SECOND).isPremiumExpired() }
+    fun nowAsLongTest() = assertTrue {
+        Clock.System.now().toEpochMilliseconds() <= nowAsLong()
+    }
+
+    @Test
+    fun nowAsInstantTest() = assertTrue {
+        Clock.System.now().toEpochMilliseconds() <= nowAsInstant().toEpochMilliseconds()
+    }
+
+    @Test
+    fun isItOver() {
+        assertTrue { (nowAsLong() - DAY).isItOver() }
+        assertTrue { (nowAsLong() - SECOND).isItOver() }
+        assertFalse { (nowAsLong() + DAY).isItOver() }
+        assertFalse { (nowAsLong() + SECOND).isItOver() }
     }
 
     @Test
@@ -40,57 +48,6 @@ internal class DateUtilTest : BaseTest() {
         "09:12 20.12.2020",
         Instant.parse("2020-12-20T09:12:28Z").toDateString(TimeZone.UTC)
     )
-
-    @Test
-    fun calculatePremiumEnd() = nowAsLong().let {
-        assertEquals(
-            it.toInstant().plus(
-                DateTimePeriod(days = REWARDED_AD_PREMIUM_IN_DAYS),
-                TimeZone.currentSystemDefault()
-            ),
-            PremiumType.VIDEO.calculatePremiumEnd(it).toInstant()
-        )
-
-        assertEquals(
-            it.toInstant().plus(
-                DateTimePeriod(months = 1),
-                TimeZone.currentSystemDefault()
-            ),
-            PremiumType.MONTH.calculatePremiumEnd(it).toInstant()
-        )
-
-        assertEquals(
-            it.toInstant().plus(
-                DateTimePeriod(months = 3),
-                TimeZone.currentSystemDefault()
-            ),
-            PremiumType.QUARTER.calculatePremiumEnd(it).toInstant()
-        )
-
-        assertEquals(
-            it.toInstant().plus(
-                DateTimePeriod(months = 6),
-                TimeZone.currentSystemDefault()
-            ),
-            PremiumType.HALF_YEAR.calculatePremiumEnd(it).toInstant()
-        )
-
-        assertEquals(
-            it.toInstant().plus(
-                DateTimePeriod(years = 1),
-                TimeZone.currentSystemDefault()
-            ),
-            PremiumType.YEAR.calculatePremiumEnd(it).toInstant()
-        )
-
-        assertEquals(
-            it.toInstant().plus(
-                DateTimePeriod(years = 100),
-                TimeZone.currentSystemDefault()
-            ),
-            PremiumType.LIFE_TIME.calculatePremiumEnd(it).toInstant()
-        )
-    }
 
     @Test
     fun doubleDigits() {
