+++ client/src/commonTest/kotlin/com/oztechan/ccc/client/viewmodel/PremiumViewModelTest.kt
@@ -12,11 +12,11 @@ import com.oztechan.ccc.client.model.PremiumData
 import com.oztechan.ccc.client.model.PremiumType
 import com.oztechan.ccc.client.storage.app.AppStorage
 import com.oztechan.ccc.client.util.calculatePremiumEnd
+import com.oztechan.ccc.client.util.nowAsLong
 import com.oztechan.ccc.client.viewmodel.premium.PremiumEffect
 import com.oztechan.ccc.client.viewmodel.premium.PremiumViewModel
-import com.oztechan.ccc.common.core.infrastructure.util.DAY
-import com.oztechan.ccc.common.core.infrastructure.util.SECOND
-import com.oztechan.ccc.common.core.infrastructure.util.nowAsLong
+import com.oztechan.ccc.common.core.infrastructure.constants.DAY
+import com.oztechan.ccc.common.core.infrastructure.constants.SECOND
 import io.mockative.Mock
 import io.mockative.classOf
 import io.mockative.given
