+++ common/core/infrastructure/src/commonMain/kotlin/com/oztechan/ccc/common/core/infrastructure/constants/TimeConstants.kt
@@ -1,0 +1,4 @@
+package com.oztechan.ccc.common.core.infrastructure.constants
+
+const val SECOND: Long = 1000L
+const val DAY: Long = 24 * 60 * 60 * SECOND
