--- common/core/infrastructure/src/commonMain/kotlin/com/oztechan/ccc/common/core/infrastructure/util/DateUtil.kt
@@ -1,10 +1,0 @@
-package com.oztechan.ccc.common.core.infrastructure.util
-
-import kotlinx.datetime.Clock
-
-const val SECOND: Long = 1000L
-const val DAY: Long = 24 * 60 * 60 * SECOND
-
-fun nowAsLong() = nowAsInstant().toEpochMilliseconds()
-
-fun nowAsInstant() = Clock.System.now()
